app/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
django-celery-beat==2.3.0
django-templated-mail==1.1.1
django-timezone-field==5.0
Django==4.0.7
djangorestframework-simplejwt==4.8.0
djangorestframework==3.13.1
djoser==2.1.0
env==0.1.0
export==0.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
pip==21.3.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests-oauthlib==1.3.1
requests==2.28.1
setuptools==60.2.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
utils==1.0.1
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.1
gunicorn==20.1.0
whitenoise==6.2.0
